requirements.txt
openai~=0.26.0
Flask~=2.2.2
python-dotenv~=0.21.0
pytest~=7.2.0
requests~=2.28.1
aiogram~=2.24